move integration tests to its own folder
