+++ packages/protocol/test/integration/LibReceiptDecoder.integration.ts
